Define status data model
The model is oriented around Prometheus metrics; as such, it differs
from the corresponding model in the Espresso status API. The main
interface for querying the data is the `prometheus` function, which
returns all available metrics in the Prometheus format. Relatedly,
the interface for updating the data simply fetches a `Box<dyn Metrics>`,
which can then be shipped off to HotShot, which will register metrics
and update each one asynchronously.

There are also some query methods to support the more targeted queries
specified in the API -- things like mempool statistics, view counters,
etc. These queries are specified separately from the Prometheus metrics
in the interface, but in the implementation of this trait, they are
defined in terms of metrics. For example, the view success rate query
simply fetches the view counter metrics from the Prometheus data and
uses these to compute the success rate.

Closes #4