Specify extensibility and define APIs to support it.
One of the crucial goals of this crate is that the API modules
provided here be extensible and composable into a larger application.
This change adds documentation describing exactly what that means and
how the user can accomplish it. Many of the APIs are reworked or
extended to support this.

First step towards #5.
Fixes #12