Add an integration test for state update; populate new data from HotShot
HotShot now provides QCs for each leaf including the leaf and block
commitments, as well as the height for each leaf. These fields are
now used to populate the query data.

An integration test spawns a network of HotShot nodes, executes some
transactions, and checks that these transactions are reflected correctly
in the query data.

Closes #15