Change getproposals return type to LeafQueryData
Previously it was return BlockQueryData, but we should stick to the
convention that LeafQueryData is the canonical object that you should
mostly work with, since it includes all the metadata and signatures.
If you want the block, you can always query for it using the block
hash in the LeafQueryData.