Revert to old nixpkgs version without broken cargo-watch
cargo-watch has apparently been broken on MacOS since
4fb9c022dda6ad2b5e7943fb55b53bd3fde96516. This fix reverts only the
update to nixpkgs, so we are using a slightly older version of
nixpkgs with an older version of cargo-watch which is not broken.

We should monitor or search for a fix for the cargo-watch issue so
that we can quickly re-upgrade our nixpkgs dependency.