Use deployer as trusted aggregator account
With the upstream updates in zkevm-node they default to using the second
account for the trusted aggregator. However this does currently still not
match with how the contracts for the L1 geth image are deployed, where
the deployer account is used instead.

In order to avoid to having to customize this address when switching
between anvil and geth backends we can go back to using the deployer
account as well.