Include number of new blocks in NewBlocks event.
This allows clients to recover the block commitments which were
added by reading a range of `numBlocks` starting from `firstBlockNumber`
in the `commitments` mapping in contract storage. This is a bit
more work than simply reading the commitments from calldata, but
it works even if the HotShot contract was called from another
contract.