Fix slow L1 test
* Use different accounts for HotShot commitment task and executor
* Fix off-by-one error in nonces
* More descriptive error messages

The test still takes a very long time to run. This is because the
executor sends 1 block at a time to the L1, which is exactly the
sort of issue this test is aimed at. To finish this off I think we
need a batch verify function in the mock L2 rollup contract, which
will anyways make it a slightly more realistic example.