Deploy contracts with fast L1 even in slow L1 test
We are interested in what happens with the HotShot commitment task
and executor task when the L1 is much slower than HotShot, but we
don't need the L1 to be slow while we're just setting things up.
This change deploys the contracts with on-demand mining and then
restarts the Anvil server with a slow block time for the meat of
the test.

This brings the runtime of the slow L1 test down from over 2 minutes
to 18 seconds.