Fix race condition in sequencer::api::serve.
When cloning a HotShotHandle, it is always important to take the
clone _before_ any events you may want to process have been emitted,
because the clone won't inherit buffered events from the original.
This is a sharp edge that we really should smooth over soon.

I don't have a good way of confirming if this is the bug that was
causing the slow L1 test to fail, but the shoe fits. I'm very glad
I added those extra checks to the rollup contract, because this is
an actual bug that has gone undetected for months.