Make sequencer API server optional
In order to make both the commitment task and the API server
optional in a reasonable way, I had to finally solve the problem
of clap not allowing multiple sub-commands. The solution is a bit
hacky but I think it's fine for now: each optional module is
enabled by a normal option, which takes as an argument a string
consisting of arguments passed to that module. The option structs
for each module derive `Parser` and use `try_parse_from` to parse
this string as additional command line arguments (this also takes
environment variables into account).