Update HotShot
This is a big change as it not only updates to async HotShot from
the run_view_refactor branch, it is also the first time we have
run the query service tests with sequencing consensus.

Actually, most of the non-trivial changes stem from the latter. In
particular, the indexes by block hash and transaction hash have
been updated to only return the earliest instance in case of
duplicates. With validating consensus, we never had duplicates.
The mock types and network for testing have also been reworked
significantly.