Improve error handling in `contract_send`
While stress testing, I observed that this function would
occasionally hang, presumably in an ethers.rs call, stalling the
commitment task and in turn all updates to the L2. I'm not sure
exactly where it was hanging, as the Future impl for
`PendingTransaction` is quite complicated and doesn't have useful
logging. But this rewrite handles all error cases and at the very
least logs at DEBUG level every time it loops. It seems to fix
the problem, and if we ever run into similar problems in the future,
it will be more transparent where exactly things are getting stuck.