Make HotShot contract idempotent
The contract will now revert if called with blocks that have
already been sequenced, or not yet been sequenced. This will at
least catch errors that we have seen occur when the commitment
task gets out of sync with the contract, for example by losing
track of a transaction which actually succeeded, or due to a reorg.

Also changed the transaction monitoring to retry a few times before
a transaction is considered failed, as on Sepolia it can take some
time after submitting before a transaction appears in the mempool.