Improve resiliency of HotShot commitment task
The previous iteration used a "dead reckoning" approach, where it
streamed blocks from HotShot and unconditionally sent them to the
L1 contract without checking the contract state. This meant that
if the task got out of sync with the contract state in any way that
caused its next transaction to fail, it would continue retrying the
bad transaction forever.

There are many ways for the task to get out of sync with the contract
including,
* an L1 reorg
* a competing task winning a race
* considering a transaction failed when in reality the transaction
  was just slow, and would later succeed in modifying the contract

The latter can happen as a result of the recent change a511ce3a4ec3093cedaa318ecc36dda1e02ce543,
which attempted to improve resilience by considering a transaction
failed if it no longer appears in the mempool. However, it can be
difficult to distinguish between a transaction which has been
tossed out of the mempool and one which is simply slow to show up
in the mempool, and the latter case has been observed to cause
problems on Sepolia.

This change redesigns the commitment task to use "live reckoning",
where it syncs with the state of the contract each time it tries
to sequence new blocks, and it fetches and sequences exactly the
blocks which are indicated by the state of the contract, rather than
by its own internal state. This should make the task more resilient
and better able to recover from both external irregularities and
bugs elsewhere in the code.