Add timestamp and L1 block info to sequencer blocks
The timestamp is needed because we work in terms of block commitments,
not leaf commitments, so clients cannot authenticate the timestamp
from the leaf. Perhaps we should change this.

I had to do a pretty ugly hack to set up an L1 provider and get the
latest L1 block when proposing a new block. There are some (I think
simple) changes to HotShot APIs that would make this much cleaner.

Finally, ntoe that this is not secure. The proposer has full
discretion to choose this metadata, and safety invariants (like
monotonicity) as well as performance requirements (like using a
sufficiently recent L1 block) are not currently enforced. The plan
is to eventually hook validation code into HotShot replicas via
the metastate API.

Closes #536