Define reference objects of types for porting
The sequencer and the types it defines have encoding-level interfaces
that we will need to create implementations for in other languages
(especially Go). For testing, it is helpful to have a library of
language-agnostic reference representations of various types.

Define JSON representations of various important types, record their
hard-coded commitments, and add regression tests in case the commitment
scheme ever changes. These same tests can thus be ported to Go and
other languages to test a port of the commitment implementation to
those languages.