Move L1 interactions off critical path, to asynchronous task
Instead of querying the L1 directly to find the latest head and
finalized block each time we need to propose, we now have each
sequencer node run a background task to update an in-memory
snapshot of the L1 state each time there is a new L1 block, and the
proposer logic just reads from this snapshot.

This has numerous benefits, including
* Drastically lower Infura usage. Instead of 2 RPC calls per HotShot
  block (one to fetch the head and one to fetch the finalized block)
  we have one RPC call per sequencer node per L1 block, plus one
  subscription stream per sequencer. This is far more optimized for
  the case where L1 blocks are much slower than HotShot blocks. For
  example, with a 1s HotShot block time, 12s L1 block time, and 9
  HotShot nodes, the total RPC calls are reduced from 2/s to 0.75/s.
* Move L1 interactions off critical path. Before, failures or delays
  in Infura calls could cause a HotShot node to miss or delay a
  proposal. We suspect this may have contributed to irregular block
  production that was ultimately observed in the OP rollup. Now,
  Infura failures and delays can cause a sequencer node to propose
  with a slightly old L1 block number, but not to miss a proposal.
* Closer to the desired design. This is more similar to how the L1
  interactions _should_ work once each sequencer is running its own
  L1 light client in the background. All we have to do to make this
  design production-ready, once the ValidatedState API is ready, is
  use ValidatedState to replace the process-global L1Client with an
  L1Client in the ValidatedState.

The only potential downside I see to deploying this right now is
that it adds fairly complicated code that has not been thoroughly
tested. However, I would argue that the failures in Cortado mean
that all the local testing we did on the previous iteration was not
sufficient to thoroughly test it either, and that therefore we
might as well just deploy this code and stress test it in a realistic
setting (Sepolia). That's what testnets are for, after all.

Local testing: I have run a few unit tests, and the rest will run in
CI. I have run the local devent with both Polygon and OP rollups
deployed on the sequencer demo. End-to-end tests for both pass and
the rollups run without crashing for at least 15-30 minutes.