Abandon infinite retry loop for contract calls
A comment argued that we could retry indefinitely as long as the
transaction is pending in the mempool, because eventually it will
either be mined or dropped. While this sounds reasonable, it is
evidently not the case, as we hit an issue on Sepolia where this case
led to infinite retries. It is safer to always consider the transaction
failed after a finite number of retries.