async-ify data source traits.
Theoretically, we can handle many requests in parallel on a single
thread, using a shared lock on the state. In practice, this is not
currently possible because I/O blocks the entire executor thread,
and the trait interfaces do not allow us to use async I/O in the
backend.

This change revises the interfaces to be async, allowing for the
possibility of a backend that takes advantage of this. The current
backend does not; it still does blocking I/O operations on async
threads. It will take a substantial amount of work to take do
proper async I/O, even just by using spawn_blocking to throw the
I/O onto another thread, because the LedgerLog I/O operations borrow
the cache and thus are not 'static. We will probably need to put
the cache in an Arc<RwLock> to fix this.

The new SQL backend will be designed from scratch to take advantage
of this new async interface. The existing file system backend will
be ported to async in the future when we have more time.