Simplify status data source traits to remove code duplication
The process for extracting specific metrics from a metrics object
is the same regardless of the underlying data source. The data
source's job is just to provide the metrics object. The status
data source trait has been revised to reflect this, so the
individual metrics are now provided functions of the trait.

As a result, SqlDataSource now implements StatusDataSource, and
all tests are passing.

Closes #212
Closes #214