Add example server and status tests
Closes #244. It turns out we already had tests for all the metrics
endpoints. However, there was a bug in the test for success_rate:
it checked `success_rate > 0.0`, which is true even if `success_rate`
is infinity, which can happen when HotShot is not populating the
`current_view` metric, causing division by 0.

I added checks that success rate is not only positive but finite,
although they are commented out right now due to
EspressoSystems/hotshot#2066.

I discovered this test bug by adding an example web server and
manually inspecting the status endpoints. So now we have an
example program, too.