Round timestamps to microsecond precision
Nanosecond precision is unnecessary and can't be stored in Postgres.
The leaf timestamp is not normative (not part of the leaf hash) so
doing this rounding shouldn't break any proofs. This should all be
cleaned up anyways once we abandon the leaf timestamps in HotShot
and switch to using the application-defined timestamp in the header,
which for the sequencer application has second precision.