Remove `default` parameter from `validate_migrations`
The boolean parameter is confusing: it turns the function into one
with two responsibilities, and it is not obvious at the call site
what it means. Since this parameter is only set to true in one
place (validation in `default_migrations`) I simply removed it, so
that `validate_migrations` only does the validation which is common
to default and custom migrations, and inlined the default migrations
validation into `default_migrations`.