Run natively with process-compose
Today during debugging I felt it was a bit painful that there wasn't a
quick and dirty way to run the full setup locally. This PR adds a
process-compose file that can be used to run the demo natively (except
for go-ethereum, which we usually don't debug).

I think this may come in handy the next time we want to debug, or
profile the application. The terminal UI looks pretty cool too.