Move proposal queries to new API module
Conceptually, they do not fit in with the availability API, which
is only about immutable, idempotent views of a conceptually
infinite ledger. The proposal queries do not specify a block
number, and so their responses will change over time. This
distinction will be formalized in the type system in the next
commit.