Reconfigure availability interface to support external DA fetching
The old `AvailabilityDataSource` trait was designed to look for data in local
storage and fail if it wasn't available. This closely matches what the backend
stores (`SqlDataSource` and `FileSystemDataSource`) do, but it is not a good fit
for the frontend APIs, for a number of reasons:
* Lack of flexibility. Sometimes we do want the fail-if-not-available behavior,
  such as when we are answering a short lived HTTP request. But other times we
  don't, such as when we are serving a long-lived WebSockets restriction. Trying
  to reconcile these two modes without a clean abstraction underpinning both
  lead to messy and error-prone logic.
* External DA fetching. We would like to be able to fetch missing data from an
  external DA provider transparently. The current trait is the wrong API for
  this: if we have a reliable external DA source, requests can't actually fail
  -- they can only take longer, if we need to go fetch the data. The existing
  trait neatly captures the interface of local storage, but local storage is
  only one part of the system we are trying to build, and shoehorning a more
  robust system with external DA fetching into this restrictive interface also
  creates a mess.
* Purity. The availability API is supposed to be a view into an immutable
  blockchain. It does not make sense that queries for immutable data would be
  impure, but currently they are -- two calls to the same method with the same
  inputs at different times can return different results: one could give you a
  result, while the other tells you that the requested data does not exist!

The new trait introduces the notion of an unresolved `Fetch` request, whose
result may be available immediately or may take some time to retrieve, but will
always be available _eventually_. This lets us better capture the idea that we
are viewing an immutable blockchain, and that local storage is only an
implementation detail. At most, the presence or absence of an object in local
storage can only affect the timing of a query, not its result.

While this new trait is a much better fit for the interface we want to present
at the API level, it is actually a _worse_ fit at the storage level, where we
don't actually have the capability to go fetch something if it is missing. We
address this by introducing a new layer:
* The storage layer implements something like the old data source API, which can
  fail if data is missing or not yet available
* The data source layer owns a storage implementation _and an external fetcher_,
  and provides the immutable, infallible interface defined here

For now, we only implement the storage layer (which mostly just involves moving
the old data source implementations to new modules and changing the interfaces
slightly) and the data source layer (which is a wrapper around a storage
implementation, with a notification system implementing futures that resolve
when a fetch completes). The actual fetching from an external provider is
stubbed out, and will be implemented in the next change. For now, we should have
parity with the previous implementation, but now with an architecture that has a
clearly marked hole where fetching will go.