Refactor notifier to avoid unsound AtomicOption
For once, cargo audit caught a real vulnerability that is actually
within my power to fix. Woo!

The new design uses AtomicBool instead of AtomicOption, and IMO it
is simpler anyways. Before, the AtomicOption<OneshotSender<T>> was
doing double duty: we needed an `Option` because `notify` consumes
the sender and takes it from the subscriber (the subscriber itself
is only dropped later, when we garbage collect (this can be done
much cleaner if and when Vec::drain_filter stabilizes)). We then
(ab)used this notion of taking the sender to close the channel in
order to close the channel and trigger garbage collection upon the
receiver being dropped. The receiver can be dropped from another
thread, hence the atomic part.

The new design separates these two concerns. We use a plain option
for taking the sender when we notify it, taking advantage of the
fact that `notify` has a mutable reference to the subscriber. We
then only need a boolean to alert us when the receiver has been
dropped, hence AtomicBool. The subscriber is considered closed when
_either_ the sender is `None` (closed by `notify`) or the AtomicBool
is `true` (closed by receiver).