Make persistence options independent of query service
This commit restructures the configuration of persistence so it is
now a separate optional module from the query service. The user can
choose between SQL and file system persistence, as before, and it
will be used to persist data which is required for the node to run.
Currently, this is just the network config file from the orchestrator.
If no persistence is chosen, it defaults to the local file system.

There is now only a single command to start the query service, no
more `query-fs` and `query-sql`. If enabled it will use whatever
persistence was previously chosen.

Interface changes:
* Sequencer command line requires `query`, not `query-fs` or `query-sql`
* Sequencer command line requires `storage-sql` if you want to use SQL
* There is no longer an option to run the commitment task with the sequencer
  (this was tech debt anyway, and it was easier to remove it than fix it)
* `ESPRESSO_SEQUENCER_CONFIG_PATH` is gone. The config file will be
  persisted in whatever persistence is chosen (`ESPRESSO_SEQUENCER_CONFIG_PATH`
  if using the file system)

Closes #804
Closes #886