Refactor active_fetch to make it fire-and-forget
The implementation of fetching will need to take a mutable lock on
a HashSet in order to check if the requested resource is already
being fetched. Since a fetch can be triggered by a read-only
request, we shouldn't take a mutable lock on the critical path.
Thus, this locking needs to happen in the task which is spawned to
handle the fetch.

Unfortunately, since we now don't even check whether we are going
to do a fetch until we have spawned the task, we have a case where
the spawned task can return `None` instead of returning a fetched
resources. We thus have _two_ ways `active_fetch` can fail: if the
requested resource is not known to exist (fails before spawning a
task) or if the resource is already being fetched (fails from the
spawned task).

To avoid nested Options or other complex interfaces, we simply make
active_fetch take a callback, which it will call if and only if a
resource is fetched successfully. This makes the spawned task truly
fire-and-forget: we don't need to await it and do any final processing,
since this will happen in the task itself via the provided callback.