Fix missed block notification bug
Since payloads are not necessarily unique, we would previously only
fetch a given payload once at a time, even if it belonged to
multiple blocks being fetched simultaneously. This is actually nice,
except it means that we only generate a notification for one of the
blocks: whichever one first triggered the payload to be fetched.
This meant that fetches for the other block would never complete.

The simple fix to this problem is to index requests for payloads by
block height. This means we trigger a separate fetch for each
distinct block, which ensures we don't miss any notifications. But
it also means we may request the same payload twice at the same time.

A better fix would be to completely decouple payloads from headers,
treating them as two separate resources which are only combined into
a block at the very end. Then we could have notifications for _payloads_,
and share not only the fetches but also the notifications across
multiple distinct blocks.

See #361