Fix another missed notification bug
Sometimes we fetch the same resource for two different reasons,
e.g. fetching a leaf for its own sake and simultaneously for its
corresponding block. In this case, only one of the callbacks would
run, so e.g. the leaf callback would run and we would never get
notified to fetch the block.

This change fixes the issue by allowing multiple callbacks to be
registered in a fetching task. Callbacks now have a concept of
equality and ordering, so each distinct callback is executed exactly
once, and callbacks are executed in a logical order (e.g. store leaf
before fetching block).

This also cleans up the workaround from the last commit: we are once
again able to fetch payloads only once when simultaneously fetching
multiple blocks with the same payload. We simply run one _callback_
for each block after fetching the payload.