Use upserts to ensure failsafe, idempotent SQL insertions
With the highly concurrent DA fetching implementation, it is
impossible to prevent the same data occasionally being fetched
twice simultaneously:
* We check the database and don't have an object
* A previous request for the same object completes and stores it
* We check that there are no fetches in progress, and there aren't,
  so we spawn a new one

Preventing this race condition would require either of the following
patterns, both with problems:
* Lock the set of in-progress requests before checking the database,
  so that if the object is not in the database, we can immediately
  spawn a request without releasing the lock, preventing other
  requests from starting or completing. This is problematic because
  it requires us to take an exclusive lock (either a mutex lock or an
  upgradable read lock) on the critical path of read-only requests,
  even in the common case where the object does not need to be fetched.
* Check the database both before and after locking the set of pending
  requests. This fixes the race condition in the same way as above, but
  further improves the common case because we never take an exclusive
  lock if the object is already available. However, it increases the
  number of requests made to the database, and so it's not clear that
  this is that much better than just allowing redundant fetches to
  occasionally happen.

Even if we implemented the above fixes, the highly concurrent nature
of the application means it would be difficult to reason confidently
that duplicate insertions are never possible. Therefore, I believe it
will result in more debuggable and reliable software if we simply
allow duplicate inserts at the database level. Postgres makes this easy
with its "upsert" feature, allowing us to atomically insert an object
into the database but do nothing (or update its value) if it already
exists.