Add `fetch_fee_receipts` and call it from `Header` methods (#989)
* Implement fee state

Implement fee state logic (#943). Fee state is initialized in
`Headers::genesis()` and updated in `Headers::new()`. The sparse
merkle tree representing this state (`FeeMerkleTree`) is held as a
header field for the time being (this will be replaced in a future
commit). New types `FeeAccount` and `FeeAmount` were created to
represent key and value in the tree and implement required
traits. `fee_merkle_tree_root` holding the `FeeMerkleTree::Commitment`
has been added to `Header` fields. `Header.commit()` has been updated
to include this commitment as well as that of
`block_merkle_tree_root` (which was overlooked in a previous
PR). `header.json` test fixture has also been updated.

* update reference header test expected value

Changes to headre commitment require an update to the expected value
used in tests.

---------

Co-authored-by: tbro <tbro@users.noreply.github.com>