Do not wait for peer to respond when initializing QueryServiceProvider
It is perfectly normal to have two peers pointing at each other as
providers, so waiting for the peer to come online can cause a deadlock.
It is also unnecessary: if we are connected to a peer that hasn't
started yet, fetch requests will fail, but they will get retried and
start to succeed as soon as the peer comes online. This is no different
than a peer crashing some time after startup, and we should handle it
the same.