Implement proactive fetching
This makes the availability tests work with [`NoStorage`] (with some
changes). The fetching task itself is actually fairly simple. Most
of the changes in this PR have more to do with configuration and
changes to the tests to make them work with no storage.

Aside from exempting the [`NoStorage`] tests from a few checks which
aren't actually guaranteed but which pass when we do have reliable
storage, the availability tests pass completely with no storage, with
no changes except features we had already planned to implement.

Changes to the tests to make them work with [`NoStorage`]:
* We don't enforce that a query for a non-unique payload or transaction
  returns the _first_ matching object, if the object needs to be fetched.
  We do guarantee this for objects which are stored locally, but there's
  no good way to do it when fetching the object, since the proactive fetcher
  may provide a matching object in any order.
* Moved the proposer queries to a separate test module. Since they are no
  longer part of the availability API, these queries do not support fetching
  and thus don't work with [`NoStorage`]

Closes #338