Add host to TmpDb struct
The first barrier to being able to remotely target a Docker
container is the inability to target a host other than
localhost.

This commit adds host to the TmpDb struct, and utilizes it
to populate the generated configuration. This will allow for
the client that consumes the Config to be able to target
a different host other than "localhost".

Add derived hostname for docker host

The second barrier is to actually populate the now
configurable Hostname.  We require the ability to specify
the hostname of the remote docker instance.

In order to fulfill this requirement I've opted to specify
the remote docker hostname via an environment variable
named "DOCKER_HOSTNAME".  We don't want to use
"DOCKER_HOST" as it's already in use by `docker` itself
for specifying a different target.  In this case
"DOCKER_HOSTNAME" is expected to just be a string of
the hostname itself.  It should even be able to be an ip
address if one should desire it to be.

Fixes Add Docker remote host considerations for tests #389