Make FetchingDataSource `new` function private
This function was only added as a nice internal place to put the
setup logic, but it can be confusing when exposed publicly because
it conflicts with the `Builder` mechanism of constructing the data
source.

For now, hiding `new` and _only_ allowing the data source to be
constructed via a builder (or the higher-level storage-specific
constructors) is a better interface.