Add `/node/sync-status` endpoint to get information about missing data
This endpoint can be used to determine if the database is fully
synced and, if not, how much data is missing. This is useful for
alerts as well as for the block explorer to be able to inform the
user when aggregate queries may be inaccurate due to syncing in
progress.

The query used to count missing data is not indexed or optimized,
but I believe it can be made efficient in the future:
* The `count(*)` to determine the number of leaves we have can be
  slow when the `leaf` table is large, because Postgres uses a full
  table scan to count. This could be made more efficient a number of
  ways, including:
  - Using estimated counts instead of exact counts
  - Storing in the database the latest count and using triggers to
    update it on inserts/updates
  - Inserting placeholder rows with NULL columns for leaves we don't
    have. This would ensure that there is a row for each possible leaf
    height, and the presence or absence of the leaf becomese a boolean
    property we can express on that row. Instead of counting over the
    entire table, we only need to count `WHERE data IS NULL` or similar,
    and we can use an index to efficiently process that filter.
* The `WHERE data IS NULL` filter for counting missing payloads is not
  indexed. We will add an index there after we test with a large enough
  database to be confident that Postgres is using the index and it is
  improving performance.

These optimizations are intentionally left as future work, for after we
have run this for some time and with large enough datasets that we can
really see the effects of various optimization strategies.

Of course, another approach to optimization would be to keep the count
of missing blocks and leaves in memory, and maintain it whenever we
`insert_block` or `insert_leaf` -- dead reckoning. This would probably
be the most efficient approach, but since this endpoint is designed to
be used for alerting and detecting low-level incidents like bugs in the
storage layer or failures in RDS, in many of the situations we care about,
the actual state of the database could end up diverging from our dead
reckoning estimate, and we might miss an incident. Therefore, I think
the endpoint is much more useful if it is running some meaningful
query over the actual data.

Closes #362