Add Block Summaries Endpoints
For the sake of the Block Explorer, as well as potential future utility we
find ourselves in need of a data structure that can describe aspects of
a Block without transmitting the full contents of a Block.  Currently, we
would like the ability to retrieve a Summary of a Block that indicates the
properties of a Block, as well as a summary of how many transactions
are contained within the Block.  In addition we find ourselves needing
to be able to grab a range of these Block Summaries for Pagination
and comparison purposes.

Add BlockSummaryQueryData type

The BlockSummaryQueryData defines the data layout of the Block
Summary endpoints.

Add From trait for BlockQuerySummaryData from BlockQueryData

Since the BlockQuerySummaryData will eventually be able to be
generated entirely from the BlockQueryData, it makes sense to be
able to actively convert to the Summary type given a Block itself.

Add get_block_summary endpoint

This endpoint can retrieve a single block_summary for a given
block height to indicate its position within the ledger.

Add get_block_summary_range endpoint

This endpoint allows us to retrieve a a range of block summaries
so we can grab a collection of them for direct reference.