Implement summary endpoints
We now have endpoints to get the total number of transactions
sequenced and the total size of all data sequenced. This is useful
for the block explorer, internal reporting, and marketing.

I also deleted the `UpdateNodeData` trait because it was a headache
to maintain and didn't really accomplish anything that the
`UpdateAvailabilityData` trait didn't already do. In fact, it just
made it confusing what trait was supposed to do what. It is now
explicitly documented that the node module uses the same underlying
data as the availability module.

Closes #342