Break `data_source/fetching.rs` into multiple files
It was getting unwieldy with all the different `Fetchable` implementations
and request types and what not. Was hard to see what need to be added or
modified when adding a new fetchable type.

Now each independent type has its own file, and hopefully it is much
more obvious where we need to change things if we are adding or changing
a type.