Add builder info and fee (#1064)
Obtain `fee_info` from `InstanceState` and add it  to `Header`. 

Wallet is derived from mnemonic passed in as a CLI option and ultimately its inner `Address` is added to `Header` (as `fee_info.account`. `Header` is signed by this account and the signature is validated in `Header.validate_and_appy_header()`. 

---------

Co-authored-by: tbro <tbro@users.noreply.github.com>
Co-authored-by: Jeb Bearer <jeb.bearer@gmail.com>