Refactor sequencer initialization, load saved leaf from storage.
The primary problem being solved here is: upon restart, load the
saved leaf from storage, instead of taking it on the command line.
This makes restart less error prone, since the node operator doesn't
have to worry about finding the last decided leaf, and ensuring it
does not cause their node to double vote (if the node was ahead and
had already voted on a future view). Instead, each node knows at all
times the highest leaf and the highest view in which it has voted.

This also simplifies startup because we don't have to fetch the leaf
data from a peer, we have it in storage even if we aren't an archival
node.

However. When I went to refactor initialization to load things from
storage, things got very messy. Basically, our initialization code has
been a dumpster fire for a long time, with all sorts of different types
being used for more or less the same thing (e.g. `SequencerNode`,
`SequencerContext`, and various ad hoc uses of `SystemContextHandle`),
background tasks being spawned all over the place and not shut down
properly, and lots of complicated code detached from any type (e.g.
`init_hotshot`) which leads to functions with huge numbers of arguments
since we don't have a struct encapsulating information needed to start
up a node.

This commit attempts to rectify the situation by standardizing on
`SequencerContext` everywhere we have a sequencer, and making sure
this context encapsulates all of the relevant information, including
background tasks. The setup logic of `init_hotshot` is also moved into
the constructor for this type. A few other related cleanups include
* An encapsulated type for the state signature stuff
* `TestConfig` and `TestNetwork` to make the startup _and cleanup_ of
  tests much cleaner