Use VID common to compute namespace proofs
This uses the new query service feature to fetch the VID common data
for the block of interest. This eliminates the need to run VID disperse
in the namespace proof endpoint. It also ensures we construct the VID
scheme with the correct number of storage nodes.

Note that for now this endpoint will fail for blocks 0 and 1, due to
HotShot not running VID for these blocks. The payloads will always be
empty, and generally rollups will be deployed after these blocks, so
this shouldn't be a problem. It will be fixed in the next release of
the query service.

Closes #1093