Fix VID genesis edge case
Currently we have an edge case where VID data will always be
unavailable for the first 2 views, since HotShot does not do VID
for these views. This is really messy and annoying for clients, as
we can see in our tests, and tests in the sequencer.

Luckily, the payload in these views is well-known (it's the genesis
payload) and the number of storage nodes is arbitrary, since we
aren't actually dispersing anything. Thus, we have all the information
we need to do VID ourselves and present a more reliable interface to
our clients.