Implement timestamp window queries
These were previously implemented as extensions in the sequencer,
but now that we have an interface for getting the timestamp from
headers (needed for pruning) it is possible to implement them here.
I think it is good to move as much functionality as possible into
the core query service rather than extensions.